Fixed Health Bar Script (Sort of)
Health actually changes now, but the entity manager assigned to each health bar gets unassigned sometimes for some reason. But you can the add the prefab to whatever scene is being used, unless it's not wanted there for this build, and plug it them back in if that happens.

The biggest problem at the moment is that since the characters' health isn't set until they are first used, their health bars empty at the start.